+++ crates/guest-write-file/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## [0.1.129](https://github.com/vm0-ai/vm0/compare/guest-write-file-v0.1.128...guest-write-file-v0.1.129) (2026-09-03)
+
 ## [0.1.128](https://github.com/vm0-ai/vm0/compare/guest-write-file-v0.1.127...guest-write-file-v0.1.128) (2026-09-03)
 
 ## [0.1.127](https://github.com/vm0-ai/vm0/compare/guest-write-file-v0.1.126...guest-write-file-v0.1.127) (2026-09-03)